--- src/app/(portafolio)/Contact/page.tsx
@@ -1,7 +1,0 @@
-export default function Contact() {
-  return (
-  <>
-    contact
-  </>
-  );
-}
